Drag and drop partial solution
you need to put your mouse over the slot for it to go back in that slot. currently they can interchangeably go between the crafting panel and the inventory panel. the problem is that if you leave them in the crafting panel, exit crafting, pick up another item - everything goes back to inventory and there is a duplication of items.